requirements.txt
gradio==5.6.0
openpyxl==3.1.2
peft==0.7.1
transformers[torch]==4.46.2
pyyaml==6.0.1
xlrd==2.0.1
python-docx==1.1.2
mammoth==1.8.0
docx2md==1.0.1
python-pptx==1.0.2
markdown==3.7
python-dotenv
beautifulsoup4==4.12.3